internal/e2e: refactor e2e tests
Now we have a the full set of xDS e2e tests, refactor file names to
match.
