+++ internal/e2e/cds_test.go
@@ -14,15 +14,53 @@
 package e2e
 
 import (
+	"context"
 	"testing"
 	"time"
 
 	v2 "github.com/envoyproxy/go-control-plane/api"
 	"github.com/gogo/protobuf/types"
 	cgrpc "github.com/heptio/contour/internal/grpc"
+	"google.golang.org/grpc"
 	"k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/util/intstr"
 )
+
+// heptio/contour#186
+// Cluster.ServiceName and ClusterLoadAssignment.ClusterName should not be truncated.
+func TestClusterLongServiceName(t *testing.T) {
+	rh, cc, done := setup(t)
+	defer done()
+
+	rh.OnAdd(service(
+		"kuard",
+		"kbujbkuhdod66gjdmwmijz8xzgsx1nkfbrloezdjiulquzk4x3p0nnvpzi8r",
+		v1.ServicePort{
+			Protocol:   "TCP",
+			Port:       8080,
+			TargetPort: intstr.FromInt(8080),
+		},
+	))
+
+	// check that it's been translated correctly.
+	assertEqual(t, &v2.DiscoveryResponse{
+		VersionInfo: "0",
+		Resources: []*types.Any{
+			any(t, &v2.Cluster{
+				Name: "kuard/kbujbkuhdod66-edfcfc/8080",
+				Type: v2.Cluster_EDS,
+				EdsClusterConfig: &v2.Cluster_EdsClusterConfig{
+					EdsConfig:   apiconfigsource("xds_cluster"), // hard coded by initconfig
+					ServiceName: "kuard/kbujbkuhdod66gjdmwmijz8xzgsx1nkfbrloezdjiulquzk4x3p0nnvpzi8r/8080",
+				},
+				ConnectTimeout: 250 * time.Millisecond,
+				LbPolicy:       v2.Cluster_ROUND_ROBIN,
+			}),
+		},
+		TypeUrl: cgrpc.ClusterType,
+		Nonce:   "0",
+	}, fetchCDS(t, cc))
+}
 
 // Test adding, updating, and removing a service
 // doesn't leave turds in the CDS cache.
@@ -365,4 +403,28 @@ func TestClusterRenameUpdateDelete(t *testing.T) {
 		TypeUrl:     cgrpc.ClusterType,
 		Nonce:       "0",
 	}, fetchCDS(t, cc))
+}
+
+func fetchCDS(t *testing.T, cc *grpc.ClientConn) *v2.DiscoveryResponse {
+	t.Helper()
+	rds := v2.NewClusterDiscoveryServiceClient(cc)
+	ctx := context.Background()
+	ctx, cancel := context.WithTimeout(ctx, 100*time.Millisecond)
+	defer cancel()
+	resp, err := rds.FetchClusters(ctx, new(v2.DiscoveryRequest))
+	if err != nil {
+		t.Fatal(err)
+	}
+	return resp
+}
+
+func apiconfigsource(clusters ...string) *v2.ConfigSource {
+	return &v2.ConfigSource{
+		ConfigSourceSpecifier: &v2.ConfigSource_ApiConfigSource{
+			ApiConfigSource: &v2.ApiConfigSource{
+				ApiType:      v2.ApiConfigSource_GRPC,
+				ClusterNames: clusters,
+			},
+		},
+	}
 }
