+++ internal/e2e/rds_test.go
@@ -20,7 +20,6 @@ import (
 	"time"
 
 	v2 "github.com/envoyproxy/go-control-plane/api"
-	"github.com/gogo/protobuf/proto"
 	"github.com/gogo/protobuf/types"
 	cgrpc "github.com/heptio/contour/internal/grpc"
 	"google.golang.org/grpc"
@@ -202,42 +201,6 @@ func TestIngressPathRouteWithoutHost(t *testing.T) {
 	}, fetchRDS(t, cc))
 }
 
-// heptio/contour#186
-// Cluster.ServiceName and ClusterLoadAssignment.ClusterName should not be truncated.
-func TestClusterLongServiceName(t *testing.T) {
-	rh, cc, done := setup(t)
-	defer done()
-
-	rh.OnAdd(service(
-		"kuard",
-		"kbujbkuhdod66gjdmwmijz8xzgsx1nkfbrloezdjiulquzk4x3p0nnvpzi8r",
-		v1.ServicePort{
-			Protocol:   "TCP",
-			Port:       8080,
-			TargetPort: intstr.FromInt(8080),
-		},
-	))
-
-	// check that it's been translated correctly.
-	assertEqual(t, &v2.DiscoveryResponse{
-		VersionInfo: "0",
-		Resources: []*types.Any{
-			any(t, &v2.Cluster{
-				Name: "kuard/kbujbkuhdod66-edfcfc/8080",
-				Type: v2.Cluster_EDS,
-				EdsClusterConfig: &v2.Cluster_EdsClusterConfig{
-					EdsConfig:   apiconfigsource("xds_cluster"), // hard coded by initconfig
-					ServiceName: "kuard/kbujbkuhdod66gjdmwmijz8xzgsx1nkfbrloezdjiulquzk4x3p0nnvpzi8r/8080",
-				},
-				ConnectTimeout: 250 * time.Millisecond,
-				LbPolicy:       v2.Cluster_ROUND_ROBIN,
-			}),
-		},
-		TypeUrl: cgrpc.ClusterType,
-		Nonce:   "0",
-	}, fetchCDS(t, cc))
-}
-
 func fetchRDS(t *testing.T, cc *grpc.ClientConn) *v2.DiscoveryResponse {
 	t.Helper()
 	rds := v2.NewRouteDiscoveryServiceClient(cc)
@@ -245,19 +208,6 @@ func fetchRDS(t *testing.T, cc *grpc.ClientConn) *v2.DiscoveryResponse {
 	ctx, cancel := context.WithTimeout(ctx, 100*time.Millisecond)
 	defer cancel()
 	resp, err := rds.FetchRoutes(ctx, new(v2.DiscoveryRequest))
-	if err != nil {
-		t.Fatal(err)
-	}
-	return resp
-}
-
-func fetchCDS(t *testing.T, cc *grpc.ClientConn) *v2.DiscoveryResponse {
-	t.Helper()
-	rds := v2.NewClusterDiscoveryServiceClient(cc)
-	ctx := context.Background()
-	ctx, cancel := context.WithTimeout(ctx, 100*time.Millisecond)
-	defer cancel()
-	resp, err := rds.FetchClusters(ctx, new(v2.DiscoveryRequest))
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -289,29 +239,6 @@ func cluster(cluster string) *v2.Route_Route {
 	}
 }
 
-func any(t *testing.T, pb proto.Message) *types.Any {
-	t.Helper()
-	any, err := types.MarshalAny(pb)
-	if err != nil {
-		t.Fatal(err)
-	}
-	return any
-}
-
-func assertEqual(t *testing.T, want, got *v2.DiscoveryResponse) {
-	t.Helper()
-	m := proto.TextMarshaler{Compact: true, ExpandAny: true}
-	a := m.Text(want)
-	b := m.Text(got)
-	if a != b {
-		m := proto.TextMarshaler{
-			Compact:   false,
-			ExpandAny: true,
-		}
-		t.Fatalf("\nexpected:\n%v\ngot:\n%v", m.Text(want), m.Text(got))
-	}
-}
-
 func service(ns, name string, ports ...v1.ServicePort) *v1.Service {
 	return &v1.Service{
 		ObjectMeta: metav1.ObjectMeta{
@@ -320,17 +247,6 @@ func service(ns, name string, ports ...v1.ServicePort) *v1.Service {
 		},
 		Spec: v1.ServiceSpec{
 			Ports: ports,
-		},
-	}
-}
-
-func apiconfigsource(clusters ...string) *v2.ConfigSource {
-	return &v2.ConfigSource{
-		ConfigSourceSpecifier: &v2.ConfigSource_ApiConfigSource{
-			ApiConfigSource: &v2.ApiConfigSource{
-				ApiType:      v2.ApiConfigSource_GRPC,
-				ClusterNames: clusters,
-			},
 		},
 	}
 }
